mistakenly moved fragment view to viewmodel (reverted)
