+++ app/src/main/java/com/searoth/template/util/Conversion.kt
@@ -1,0 +1,6 @@
+package com.searoth.template.util
+
+/**
+ * Created by yusuf on 10/03/2018.
+ */
+fun convertToCurrency(countryCurrency: String, value: Double): String = "$countryCurrency $value"
