+++ app/src/main/res/layout/fragment_home.xml
@@ -2,7 +2,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.searoth.template.viewmodel.main.fragment.HomeFragment">
+    tools:context="com.searoth.template.view.main.fragment.HomeFragment">
 
     <android.support.constraint.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
         xmlns:app="http://schemas.android.com/apk/res-auto"
